#hotfix-74 queries selection dropdown ui design change
