+++ applications/virtual-fly-brain/client/src/shared/subHeader/QueriesSelection/index.js
@@ -1,15 +1,15 @@
-import { Box, Typography, Button, Popper, List, ListItem, ListItemButton, ListItemText } from "@mui/material";
 import React from "react";
-import { AngleRight, ChevronDown, CleaningServices, Delete, ChevronUp } from "../../../icons";
+import { Box, Typography, Button } from "@mui/material";
+import { AngleRight, CleaningServices } from "../../../icons";
+import { QueriesSelectionDropdown } from "./QueriesSelectionDropdown";
 import vars from "../../../theme/variables";
 
 const queryBuilderDatasourceConfig = require('../../../components/configuration/VFBSearchBuilder/queryBuilderConfiguration').queryBuilderDatasourceConfig;
 
-const { searchHeadingColor, searchBoxBg, primaryBg, whiteColor, queryBorderColor, btnDisabledColor } = vars;
+const { searchHeadingColor, primaryBg, btnDisabledColor } = vars;
 
 export const QueriesSelection = ({ checkResults, handleQueryDeletion, recentSearch }) => {
   const [searchQueries, setSearchQueries] = React.useState([]);
-  const [popoverAnchorEl, setPopoverAnchorEl] = React.useState(null);
   const [selectedOption, setSelectedOption] = React.useState({ count : 0});
   const [selectedQueryIndex, setSelectedQueryIndex] = React.useState("");
   React.useEffect(() => {
@@ -24,29 +24,7 @@ export const QueriesSelection = ({ checkResults, handleQueryDeletion, recentSear
     let updateSelection = selectedOption;
     setSelectedOption(updateSelection)
   }
-  const popoverHandleClick = (event) => {
-    setSelectedQueryIndex(event.target.parentElement.id || event.target.id)
-    setPopoverAnchorEl(popoverAnchorEl ? null : event.target.parentElement.parentElement);
-  };
 
-  const handleSelect = (option, query) => {
-    let count = 0;
-    if ( query.queries?.Examples && !selectedOption[query.short_form]) {
-      count = Object.keys(query.queries?.Examples)?.length;
-    }
-    Object.keys(selectedOption)?.forEach( o => {
-      if ( typeof selectedOption[o] === 'object' ) {
-        count = count + ( selectedOption[o].count || 0 )
-      } 
-    })
-    let updatedSelectedOption = {...selectedOption, [query.short_form]: option, count : count};
-    updatedSelectedOption[query.short_form].count =  Object.keys(query.queries?.Examples)?.length || 0;
-    setSelectedOption(updatedSelectedOption)
-    setPopoverAnchorEl(null);
-  };
-
-  const popoverOpen = Boolean(popoverAnchorEl);
-  const id = popoverOpen ? 'simple-popover' : undefined;
   return (
     <Box sx={{
       py: '1rem',
@@ -96,126 +74,15 @@ export const QueriesSelection = ({ checkResults, handleQueryDeletion, recentSear
         rowGap={2}
       >
         { searchQueries?.map((option, index) =>
-        <Box
-          display='flex'
-          key={option.label}
-          columnGap={1}
-        >
-          <Button
-            sx={{
-              flexShrink: 0,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              width: '1.5rem',
-              height: '1.5rem',
-              p: 0,
-              borderRadius: 1,
-              background: searchBoxBg,
-              minWidth: '0.0625rem',
-
-              '&:hover': {
-                background: searchBoxBg,
-              }
-            }}
-            key={option.short_form}
-            id={option.label}
-            onClick={deleteQuery}
-          >
-            <Delete size={12} />
-          </Button>
-
-          <Box
-            borderRadius={1}
-            display='flex'
-            sx={{
-              width: 'calc(100% - 2rem)',
-              background: searchBoxBg,
-              border: popoverOpen ? "1px solid #0AB7FE" : "none"
-            }}
-          >
-            <Box
-              width={1}
-              flexGrow={1}
-              display='flex'
-            >
-              
-              {selectedOption[option.short_form] ? 
-              (<Typography sx={{
-                width: 'calc(100% - 1.875rem)',
-                px: 1,
-                alignSelf: 'center',
-                flexGrow: 1,
-                fontSize: '0.75rem',
-                lineHeight: '133%',
-                color: whiteColor,
-                whiteSpace: 'nowrap',
-                overflow: 'hidden',
-                textOverflow: 'ellipsis'
-              }}>
-                { selectedOption[option.short_form].label }
-              </Typography>)
-              :
-              (<Typography sx={{
-                width: 'calc(100% - 1.875rem)',
-                px: 1,
-                alignSelf: 'center',
-                flexGrow: 1,
-                fontSize: '0.75rem',
-                lineHeight: '133%',
-                color: whiteColor,
-                whiteSpace: 'nowrap',
-                overflow: 'hidden',
-                textOverflow: 'ellipsis'
-              }}>
-               Select query for { option.label }
-              </Typography>) }
-              <Button
-                aria-describedby={id}
-                id={option.short_form}
-                onClick={popoverHandleClick}
-                sx={{
-                  borderLeft: `0.0625rem solid ${queryBorderColor}`,
-                  minWidth: '0.0625rem',
-                  width: '1.875rem',
-                  height: '100%',
-                  p: 0,
-
-                  '&:hover': {
-                    background: 'transparent'
-                  }
-                }}
-              >
-                {popoverOpen ? <ChevronUp/> : <ChevronDown/>}
-              </Button>
-
-              <Popper
-                placement='bottom-start'
-                id={id}
-                open={popoverOpen}
-                anchorEl={popoverAnchorEl}
-                sx={{
-                  marginTop: "0.5px !important"
-                }}
-              >
-
-                <List>
-                  {!option.queries.Queries.length && <ListItem>
-                    <ListItemButton>
-                      <ListItemText primary={`No query for ${option.label}`}/>
-                    </ListItemButton>
-                  </ListItem>
-                  }
-                  { option.queries?.Queries?.map((query, index) => (selectedQueryIndex === option.short_form || selectedQueryIndex === "" )&& (<ListItem key={query.short_form+index}>
-                    <ListItemButton onClick={() => handleSelect(query, option)}>
-                      <ListItemText primary={query.label} />
-                    </ListItemButton>
-                  </ListItem>) )}
-                </List>
-              </Popper>
-            </Box>
-          </Box>
-        </Box>
+        <QueriesSelectionDropdown
+          key={index}
+          option={option}
+          selectedOption={selectedOption}
+          setSelectedOption={setSelectedOption}
+          selectedQueryIndex={selectedQueryIndex}
+          setSelectedQueryIndex={setSelectedQueryIndex}
+          deleteQuery={deleteQuery}
+        />
         )}
         <Box
           display='flex'
